Add HTTP/3 example (Example 2)
The traefik container serves HTTP/3 both to the
internet and to the custom network.
systemd creates the sockets for socket activation.
Those sockets are inherited by traefik and used for
serving the internet.
Traefik creates the sockets on the custom network.
Those sockets are used by traefik for accepting
connections from containers on the custom network.

Signed-off-by: Erik Sjölund <erik.sjolund@gmail.com>